hello

+++ include/Az3D/Material.hpp
@@ -9,7 +9,7 @@ namespace Az3D {
     struct Material {
         // Custom material properties exclusive to my game engine
         glm::vec4 multColor{1.0f, 1.0f, 1.0f, 1.0f};
-        int diffTxtr = 0; // Albedo/Diffuse map
+        size_t diffTxtr = 0; // Albedo/Diffuse map
     };
